Remove `package` Makefile rule
It does not seem like it's a good idea to automate the building of this
package using py2dsc, etc. The files produced should really
custom-editted and versioned themselves. For example, the automated
building cannot correctly know the debian dependencies.

The `package` Makefile was removed and instructions for building were
instead put inside the PACKAGING file.